requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==3.2.9
django-cachalot==2.5.1
django-extensions==3.2.1
django-versatileimagefield==2.2
funcy==1.17
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.3.0
promise==2.3
psycopg2-binary==2.9.5
PyJWT==2.5.0
python-magic==0.4.27
pytz==2022.2.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
